back/requirements.txt
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
requests==2.24.0
